\name{dbConnectionInfo.PgSQL.conn}
\alias{dbConnectionInfo.PgSQL.conn}
\title{Return connection information}
\description{
  \code{dbConnectionInfo} returns information about a connection.
}
\usage{
\method{dbConnectionInfo}{PgSQL.conn}(conn)
}
\arguments{
  \item{conn}{A connection object}
}
\value{
  \item{status}{Connection status; zero indicates valid connection}
  \item{last.message}{Last message generated by the connection}
  \item{database.name}{The database name}
  \item{host.name}{The database host name}
  \item{options}{Additional connection options}
  \item{user.name}{The user name}
  \item{password}{The password used to connect}
  \item{port}{The port number}
  \item{tty}{The tty associated with the connection}
  \item{socket}{The socket number}
  \item{client.encoding}{Binary encoding}
  \item{backend.pid}{The PID of the backend process}
}
\references{\url{http://rdbi.sourceforge.net}}
\author{Timothy H. Keitt}

\seealso{\code{\link{dbConnect}}}

\keyword{data}
