\name{dbResultInfo.PgSQL.result}
\alias{dbResultInfo.PgSQL.result}
\title{Get result information}
\description{
  \code{dbResultInfo} return information about a result object.
}
\usage{
\method{dbResultInfo}{PgSQL.result}(result)
}
\arguments{
  \item{result}{A result object}
}
\value{
  \item{status}{Result status code}
  \item{status.string}{Result status string}
  \item{error.message}{Error message generated by query}
  \item{rows}{Number of rows in the result}
  \item{columns}{Number of columns in the result}
  \item{is.binary}{If true, result is binary}
  \item{command.response}{Response from backend}
  \item{command.tuples}{Not sure what this is}
}
\references{\url{http://rdbi.sourceforge.net/}}
\author{Timothy H. Keitt}

\seealso{\code{\link{dbSendQuery}}}

\keyword{data}

