\name{Internal functions}
\alias{format.null.values}
\alias{make.db.names}
\alias{rpgsql.cast.values}
\alias{rpgsql.cast.values.1042}
\alias{rpgsql.cast.values.1043}
\alias{rpgsql.cast.values.1082}
\alias{rpgsql.cast.values.1083}
\alias{rpgsql.cast.values.25}
\alias{rpgsql.cast.values.default}
\alias{rpgsql.data.type}
\alias{rpgsql.data.type.dates}
\alias{rpgsql.data.type.default}
\alias{rpgsql.data.type.times}
\alias{rpgsql.format.values}
\alias{rpgsql.format.values.dates}
\alias{rpgsql.format.values.default}
\alias{rpgsql.format.values.times}
\title{Internal function}
\description{
  These functions are used internally.
}
\usage{
}
\arguments{
}
\references{\url{http://rdbi.sourceforge.net/}}
\author{Timothy H. Keitt}
\keyword{internal}
